Add filename generated in controls to viewport
Add a new method (`update_state`) to the main app to allow the global state
of the application to be updated. This allows communication between different
components of the application (like `Viewport`, `Controls`, etc...).

`update_state` is used to communicate to the viewport the new image file that
should be displayed so that it can perform the update accordingly.

Also move the code to update the viewport image into it's own method to
facilitate being updated by other components.