Add remove borders button and functionality
Add a button to toggle the borders from all img_windows on/off. This 'freezes'
the windows in place however also free's up space on screen when multiple
image windows are in use.

Also fix a bug where image windows weren't being added to the `img_windows`
list when being created due to an uninitialised list. Now the list is
initialised in `__init__()` and guaranteed to exist when used in other
functions.