Fix zoom out not returning to the same place when zooming in
The previous zoom out equation was incorrect as it would not be able to scale
the image back up when zooming in. An illustration of this incorrect equation
is given:

500 * 0.9 = 450
450 * 1.1 = 495

The new equation is not perfect as the canvas has to store an int so the float
gets truncated however it is good enough to not be noticable for now.