Change image filters to increase performance
We were previously using the LANCZOS image filter to render images when
resizing upon scrolling/zooming in/out as well as when the canvas/window was
resized. This is a high quality image filter however it caused a lot of lag
when lots of events were occuring at once.

Change the image filters when resizing events occur to lower quality filters
to speed up processing in event handlers.