Add video file structural support
Add the basic structural support to support video files. Files are given
types inside the viewport and controls are added so that the user can choose
to select random video files in a similar way to image files.

This does not include support to play files which will be added in a seperate
commit.