Refactor file logic code
This is a large commit that simplifies the logical control flow of the whole
program. It introduces several new code improvements however does not alter
the program or any features in a user noticeable way. The improvements are:

- Remove file (image or video) specific code from viewport.py and controls.py.
- Generalise file logic into file.py which specialises into image_file.py and
  video_file.py. viewport.py and controls.py only handle the generalised
  file.py.
- Move image_window.py and video_window.py into it's own folder/directory.
  This clearly separates these modules and allows for the possibility of
  a more generalised window.py to be introduced in future.

This commit massively simplifies the codebase by encapsulating the file
specific logic and seperating it from the user interface logic.