Replace custom thumbnail resizing with PIL function
scale_to_canvas() was originally made as a solution to scale the size of an
image so that it would fit inside a given area/canvas while maintaining the
aspect ratio. While this worked, to reduce the code complexity we can replace
it with a function in PIL that was designed for that purpose:
Image.thumbnail().